refactor: reutilizar ItemTarea y adaptar TareasPorEstado para vista de sprint

+++ src/components/ui/ItemSprint/ItemSprint.tsx
@@ -5,6 +5,7 @@ import { IoEyeSharp } from "react-icons/io5";
 import { FaPen, FaTrashAlt } from "react-icons/fa";
 import { useSprints } from "../../../hooks/useSprints";
 import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 
 interface ItemSprint {
     sprint: ISprint;
@@ -13,9 +14,9 @@ interface ItemSprint {
 }
 
 export const ItemSprint: FC<ItemSprint> = ({ sprint, handleOpenModalVer, handleOpenModalEdit }) => {
-    const navigate =useNavigate();
-    const verListaTareasDelSprint = ()=>{
-        navigate(`/sprint/:${sprint.id}`);
+    const navigate = useNavigate();
+    const verListaTareasDelSprint = () => {
+        navigate(`/sprint/${sprint._id}`);
     }
     const { eliminarSprint } = useSprints();
 
@@ -23,8 +24,20 @@ export const ItemSprint: FC<ItemSprint> = ({ sprint, handleOpenModalVer, handleO
         handleOpenModalEdit(sprint);
     };
 
-    const eliminarSprintt = (id: string) => {
-        eliminarSprint(id);
+    const eliminarSprintt =  async  (id: string) => {
+        const result = await Swal.fire({
+            title: "¿Estás seguro?",
+            text: "Esta acción no se puede deshacer",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonText: "Sí, eliminar",
+            cancelButtonText: "Cancelar"
+        });
+
+        if (result.isConfirmed) {
+            await eliminarSprint(id);
+            Swal.fire("¡Eliminado!", "El sprint ha sido eliminado.", "success");
+        }
     };
 
     return (
@@ -39,7 +52,7 @@ export const ItemSprint: FC<ItemSprint> = ({ sprint, handleOpenModalVer, handleO
             <div className={styles.ItemSprint__buttons}>
                 <button onClick={verListaTareasDelSprint} style={{ backgroundColor: "#6BB0FF", color: "white", border: "none" }}><IoEyeSharp /></button>
                 <button onClick={editarSprint} style={{ backgroundColor: "#85C86D", color: "white", border: "none" }}><FaPen /></button>
-                <button onClick={() => { eliminarSprintt(sprint.id!); }} style={{ backgroundColor: "#FF6B6B", color: "white", border: "none" }}><FaTrashAlt /></button>
+                <button onClick={() => { eliminarSprintt(sprint._id!); }} style={{ backgroundColor: "#FF6B6B", color: "white", border: "none" }}><FaTrashAlt /></button>
             </div>
         </div>
     )
